requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
cryptography==41.0.0
pyOpenSSL==23.2.0
bcrypt==4.0.1
Pillow==10.0.0
Flask-Talisman==1.0.0
Flask-CSP==0.1.0
email_validator>=1.3.0